# Pick an optimal zstd encoding buffer based on uncompressed data size